+++ client-tests/test/foo.spec.ts
@@ -1,219 +1,13 @@
 import { describe, it, xit, expect } from "./index";
-
-async function setupPage(page) {
-  page.setViewportSize({ width: 1500, height: 800 });
-}
-
-type ColorModel =
-  | "noColor"
-  | "pairOrientation"
-  | "insertSize"
-  | "insertSizeAndPairOrientation"
-  | "readStrand"
-  | "firstInPairStrand";
-
-type TrackFormat =
-  | "BAM"
-  | "BED"
-  | "GENE"
-  | "MAF"
-  | "REFERENCE"
-  | "Ruler"
-  | "SEG"
-  | "VCF"
-  | "WIG";
-
-type DisplayMode = "Bar Graph" | "Heat Map";
-
-type PropType<TObj, TProp extends keyof TObj> = TObj[TProp];
-type VariantsView = "Collapsed" | "Expanded";
-type GroupMode = "default" | "firstInPair" | "pairOrientation" | "chromosomeOfMate" | "readStrand";
-enum ReadsViewMode {
-  Collapsed = 0,
-  Expanded = 1,
-  Automatic = 2
-}
-
-type ITrackStateDefinition = {
-  index: number | string; // figure out
-  alignments: boolean;
-  arrows: boolean;
-  colorMode: ColorModel;
-  coverage: boolean;
-  diffBase: boolean;
-  geneTranscript: "collapsed" | "expanded";
-  groupAutoScale: "default" | "manual" | "group";
-  groupMode: GroupMode;
-  ins_del: boolean;
-  mismatches: boolean;
-  readsViewMode: ReadsViewMode;
-  shadeByQuality: boolean;
-  softClip: boolean;
-  spliceJunctions: boolean;
-  variantsView: VariantsView;
-  viewAsPairs: boolean;
-  coverageDisplayMode: DisplayMode;
-  coverageScaleMode: 'default' | 'manual' | 'group';
-  coverageLogScale: boolean;
-  coverageScaleFrom: number;
-  coverageScaleTo: number;
-  referenceShowTranslation: boolean;
-  referenceShowForwardStrand: boolean;
-  referenceShowReverseStrand: boolean;
-  header: {
-    fontSize: string;
-  },
-  color: unknown; // @todo - figure out
-}
-
-type ITrackDefinition = {
-  bioDataItemId: string;
-  name: string;
-  projectId: string;
-  height: number;
-  format: TrackFormat;
-  isLocal: boolean;
-  state: ITrackStateDefinition;
-};
-
-type ITrack = {
-  projectIdNumber?: number;
-  b: PropType<ITrackDefinition, "bioDataItemId">;
-  h?: PropType<ITrackDefinition, "height">;
-  p: PropType<ITrackDefinition, "projectId">;
-  n?: PropType<ITrackDefinition, "name">;
-  l?: PropType<ITrackDefinition, "isLocal">;
-  f?: PropType<ITrackDefinition, "format">;
-  s?: {
-    rsfs?: PropType<ITrackStateDefinition, "referenceShowForwardStrand">;
-    rsrs?: PropType<ITrackStateDefinition, "referenceShowReverseStrand">;
-    rt?: PropType<ITrackStateDefinition, "referenceShowTranslation">;
-    a?: PropType<ITrackStateDefinition, "arrows">;
-    d?: PropType<ITrackStateDefinition, "diffBase">;
-    aa?: PropType<ITrackStateDefinition, "alignments">;
-    m?: PropType<ITrackStateDefinition, "mismatches">;
-    c?: PropType<ITrackStateDefinition, "colorMode">;
-    co?: PropType<ITrackStateDefinition, "color">;
-    i?: PropType<ITrackStateDefinition, "ins_del">;
-    i1?: PropType<ITrackStateDefinition, "index">;
-    g?: PropType<ITrackStateDefinition, "geneTranscript">;
-    v?: PropType<ITrackStateDefinition, "variantsView">;
-    c1?: PropType<ITrackStateDefinition, "coverage">;
-    g1?: PropType<ITrackStateDefinition, "groupMode">;
-    gas?: PropType<ITrackStateDefinition, "groupAutoScale">;
-    r?: PropType<ITrackStateDefinition, "readsViewMode">;
-    s1?: PropType<ITrackStateDefinition, "shadeByQuality">;
-    s2?: PropType<ITrackStateDefinition, "softClip">;
-    s3?: PropType<ITrackStateDefinition, "spliceJunctions">;
-    v1?: PropType<ITrackStateDefinition, "viewAsPairs">;
-    cdm?: PropType<ITrackStateDefinition, "coverageDisplayMode">;
-    cls?: PropType<ITrackStateDefinition, "coverageLogScale">;
-    csm?: PropType<ITrackStateDefinition, "coverageScaleMode">;
-    csf?: PropType<ITrackStateDefinition, "coverageScaleFrom">;
-    cst?: PropType<ITrackStateDefinition, "coverageScaleTo">;
-    he?: PropType<ITrackStateDefinition, "header">;
-    wigColors?: {
-      negative: number;
-      positive: number;
-    };
-  };
-};
-
-const hostname = `http://localhost:8080`;
-
-const pageObject = {
-  buttons: {
-    settings:
-      "body > ui-view > div > div.left-menu > ngb-main-toolbar > div > ul > li:nth-child(5) > ngb-main-settings > button",
-  },
-  windows: {
-    settings: {
-      container:
-        "body > div.md-dialog-container.ng-scope > md-dialog > form > md-toolbar > div > h2",
-      tabs: {
-        alignments: {
-          container:
-            "md-tabs > md-tabs-wrapper > md-tabs-canvas > md-pagination-wrapper > md-tab-item:nth-child(2) > span",
-          content: '[role="tabpanel"]',
-          checkboxes: {
-            downsampling: '[aria-label="Downsample reads"]',
-          },
-          buttons: {
-            save:
-              "body > div.md-dialog-container.ng-scope > md-dialog > form > md-dialog-actions > button.md-primary.md-button.md-ink-ripple > span",
-          },
-        },
-      },
-    },
-    main: {
-      trackSettings: {
-        groupBy: {
-          container:
-            "body > ui-view > div > div.golden-layout > ngb-golden-layout > div > div > div > div:nth-child(1) > div.lm_items > div > div > ngb-browser > ngb-tracks-view > div > div > div:nth-child(2) > ngb-track:nth-child(3) > div > div.ngb-track-header.layout-align-start-center.layout-row.flex-100 > ngb-track-settings > md-menu:nth-child(5) > button > a:nth-child(1)",
-          options: {
-            startLocation:
-              "#menu_container_11 > md-menu-content > div:nth-child(2) > md-menu-item > button > span:nth-child(1)",
-            insertSize:
-              "#menu_container_11 > md-menu-content > div:nth-child(6) > md-menu-item > button > span:nth-child(1)",
-          },
-        },
-      },
-      canvas:
-        "body > ui-view > div > div.golden-layout > ngb-golden-layout > div > div > div > div:nth-child(1) > div.lm_items > div > div > ngb-browser > ngb-tracks-view > div > div > div:nth-child(2) > ngb-track:nth-child(3) > div > div.ngb-track-renderer.js-ngb-render-container-target.flex-100 > div > canvas",
-    },
-  },
-};
-
-function buildUrl(
-  referenceId: string,
-  chromosome: string = "",
-  start: string,
-  end: string,
-  tracks: ITrack[] = []
-) {
-  return `${hostname}/#/${[referenceId, chromosome, start, end]
-    .filter((el) => el.length)
-    .join("/")}?tracks=${encodeURIComponent(JSON.stringify(tracks))}`;
-}
+import tracksFor from "./test-support/data-storage";
+import { pageObject, setupPage, buildUrl } from "./test-support/page-object";
+import { ITrack } from "./test-support/types";
 
 describe("foo", () => {
   xit("is a basic test with the page", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      { b: "DM6", p: "Fruitfly", f: "REFERENCE", projectIdNumber: 1 },
-      {
-        b: "agnts3.09-28.trim.dm606.realign.bam",
-        p: "Fruitfly",
-        f: "BAM",
-        projectIdNumber: 1,
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.bam",
-        p: "Fruitfly",
-        f: "BAM",
-        projectIdNumber: 1,
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.vcf",
-        p: "Fruitfly",
-        f: "VCF",
-        projectIdNumber: 1,
-      },
-      {
-        b: "CantonS.09-28.trim.dm606.realign.bam",
-        p: "Fruitfly",
-        f: "BAM",
-        projectIdNumber: 1,
-      },
-      {
-        b: "CantonS.09-28.trim.dm606.realign.vcf",
-        p: "Fruitfly",
-        f: "VCF",
-        projectIdNumber: 1,
-      },
-      { b: "DM6_Genes", p: "", l: true, f: "GENE" },
-    ];
+    const tracks: ITrack[] = tracksFor("is a basic test with the page");
 
     const url = buildUrl("DM6", "", "", "", tracks);
 
@@ -240,115 +34,7 @@ describe("foo", () => {
   xit("it able to test mouse movements", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      {
-        b: "agnts3.09-28.trim.dm606.realign.bam",
-        f: "BAM",
-        h: 250,
-        p: "Fruitfly",
-        s: {
-          a: true,
-          aa: true,
-          c: "noColor",
-          c1: true,
-          d: true,
-          g1: "default",
-          i: true,
-          m: true,
-          r: 1,
-          s1: false,
-          s2: true,
-          s3: false,
-          v1: false,
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-        },
-        projectIdNumber: 1,
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.bam",
-        f: "BAM",
-        h: 250,
-        p: "Fruitfly",
-        s: {
-          a: true,
-          aa: true,
-          c: "noColor",
-          c1: true,
-          d: true,
-          g1: "default",
-          i: true,
-          m: true,
-          r: 1,
-          s1: false,
-          s2: true,
-          s3: false,
-          v1: false,
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-        },
-        projectIdNumber: 1,
-      },
-      {
-        b: "DM6",
-        f: "REFERENCE",
-        h: 100,
-        p: "Fruitfly",
-        s: { rt: false, rsfs: true, rsrs: false },
-        projectIdNumber: 1,
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.vcf",
-        f: "VCF",
-        h: 70,
-        p: "Fruitfly",
-        s: { v: "Collapsed" },
-        projectIdNumber: 1,
-      },
-      {
-        b: "CantonS.09-28.trim.dm606.realign.bam",
-        f: "BAM",
-        h: 250,
-        p: "Fruitfly",
-        s: {
-          a: true,
-          aa: true,
-          c: "noColor",
-          c1: true,
-          d: true,
-          g1: "default",
-          i: true,
-          m: true,
-          r: 1,
-          s1: false,
-          s2: true,
-          s3: false,
-          v1: false,
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-        },
-        projectIdNumber: 1,
-      },
-      {
-        b: "CantonS.09-28.trim.dm606.realign.vcf",
-        f: "VCF",
-        h: 70,
-        p: "Fruitfly",
-        s: { v: "Collapsed" },
-        projectIdNumber: 1,
-      },
-      {
-        b: "DM6_Genes",
-        f: "GENE",
-        h: 158,
-        l: true,
-        p: "",
-        s: { g: "collapsed" },
-      },
-    ];
+    const tracks: ITrack[] = tracksFor("it able to test mouse movements");
 
     const url = buildUrl("DM6", "X", "1", "23542271", tracks);
 
@@ -374,32 +60,7 @@ describe("foo", () => {
   it("matches barchart graph", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      {
-        b: "h37",
-        p: "h37_data",
-        f: "REFERENCE",
-        projectIdNumber: 1,
-        h: 100,
-        s: { rt: false, rsfs: true, rsrs: false },
-      },
-      {
-        b: "sv_sample_2.bw",
-        h: 86,
-        p: "h37_data",
-        projectIdNumber: 1,
-        s: {
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-          csf: 100,
-          cst: 900,
-          wigColors: { negative: 16731471, positive: 3265094 },
-        },
-        n: "sv_sample_2.bw",
-        f: "WIG",
-      },
-    ];
+    const tracks: ITrack[] = tracksFor("matches barchart graph");
 
     const url = buildUrl("h37", "1", "150966170", "150966249", tracks);
     await page.goto(url);
@@ -414,40 +75,7 @@ describe("foo", () => {
   it("matches distribution graph", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      {
-        b: "h37",
-        p: "h37_data",
-        f: "REFERENCE",
-        projectIdNumber: 1,
-        h: 100,
-        s: { rt: false, rsfs: true, rsrs: false },
-      },
-      {
-        b: "Homo_sapiens.GRCh37.gtf.gz",
-        h: 84,
-        l: true,
-        p: "",
-        s: { g: "expanded" },
-        f: "GENE",
-      },
-      {
-        b: "sv_sample_2.bw",
-        h: 150,
-        p: "h37_data",
-        projectIdNumber: 1,
-        s: {
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-          csf: 100,
-          cst: 900,
-          wigColors: { negative: 16731471, positive: 3265094 },
-        },
-        n: "sv_sample_2.bw",
-        f: "WIG",
-      },
-    ];
+    const tracks: ITrack[] = tracksFor("matches distribution graph");
 
     const url = buildUrl("h37", "1", "150964883", "150967542", tracks);
 
@@ -463,43 +91,7 @@ describe("foo", () => {
   it("works for track renderer", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      {
-        b: "dm6",
-        p: "dm6_data",
-        f: "REFERENCE",
-        projectIdNumber: 2,
-        h: 100,
-        s: { rt: false, rsfs: true, rsrs: false },
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.bam",
-        h: 576,
-        p: "dm6_data",
-        projectIdNumber: 2,
-        s: {
-          a: true,
-          aa: true,
-          c: "insertSize",
-          c1: true,
-          d: true,
-          g1: "default",
-          he: { fontSize: "16px" },
-          i: true,
-          m: true,
-          r: 2,
-          s1: true,
-          s2: true,
-          s3: false,
-          v1: false,
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-        },
-        n: "agnX1.09-28.trim.dm606.realign.bam",
-        f: "BAM",
-      },
-    ];
+    const tracks: ITrack[] = tracksFor("works for track renderer");
 
     const url = buildUrl("dm6", "X", "12584385", "12592193", tracks);
 
@@ -556,43 +148,7 @@ describe("foo", () => {
   it("matches with disabled downsampling", async ({ page }) => {
     await setupPage(page);
 
-    const tracks: ITrack[] = [
-      {
-        b: "dm6",
-        p: "dm6_data",
-        f: "REFERENCE",
-        projectIdNumber: 2,
-        h: 100,
-        s: { rt: false, rsfs: true, rsrs: false },
-      },
-      {
-        b: "agnX1.09-28.trim.dm606.realign.bam",
-        h: 576,
-        p: "dm6_data",
-        projectIdNumber: 2,
-        s: {
-          a: true,
-          aa: true,
-          c: "insertSize",
-          c1: true,
-          d: true,
-          g1: "default",
-          he: { fontSize: "16px" },
-          i: true,
-          m: true,
-          r: 2,
-          s1: true,
-          s2: true,
-          s3: false,
-          v1: false,
-          cdm: "Bar Graph",
-          cls: false,
-          csm: "default",
-        },
-        n: "agnX1.09-28.trim.dm606.realign.bam",
-        f: "BAM",
-      },
-    ];
+    const tracks: ITrack[] = tracksFor("matches with disabled downsampling");
 
     const url = buildUrl("dm6", "X", "12584385", "12592193", tracks);
 
@@ -630,10 +186,11 @@ describe("foo", () => {
 
     // const coords = await element.boundingBox();
     const screenshot = await element.screenshot({ omitBackground: true });
-    expect(
-      screenshot
-    ).toMatchSnapshot(`ngb-track-renderer-canvas-no-threshold.png`, {
-      threshold: 0,
-    });
+    expect(screenshot).toMatchSnapshot(
+      `ngb-track-renderer-canvas-no-threshold.png`,
+      {
+        threshold: 0,
+      }
+    );
   });
 });
